mh-backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.1
